+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to
 
 #### Breaking Changes
 #### Added
-- cuda/ptx: add laneid.cu;
+- ptx: add globaltimer.cu, laneid.cu;
 - crypto: add fips subdir;
 #### Changed
 #### Deprecated